add feature to change user role, add user and guest apis, add payment integration with stripe

+++ src/main/java/com/bhavesh/roovia/service/hotel/HotelServiceImpl.java
@@ -11,21 +11,22 @@ import com.bhavesh.roovia.entity.User;
 import com.bhavesh.roovia.exception.ResourceNotFoundException;
 import com.bhavesh.roovia.exception.UnAuthorisedException;
 import com.bhavesh.roovia.repository.HotelRepository;
-//import com.bhavesh.roovia.repository.InventoryRepository;
+import com.bhavesh.roovia.repository.InventoryRepository;
 import com.bhavesh.roovia.repository.RoomRepository;
-//import com.bhavesh.roovia.service.inventory.InventoryService;
+import com.bhavesh.roovia.service.inventory.InventoryService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
-//import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.stream.Collectors;
 
-//import static com.bhavesh.roovia.util.AppUtils.getCurrentUser;
+import static com.bhavesh.roovia.util.AppUtils.getCurrentUser;
+
 
 @Service
 @Slf4j
@@ -34,9 +35,9 @@ public class HotelServiceImpl implements HotelService {
 
     private final HotelRepository hotelRepository;
     private final ModelMapper modelMapper;
-//    private final InventoryService inventoryService;
-//    private final RoomRepository roomRepository;
-//    private final InventoryRepository inventoryRepository;
+    private final InventoryService inventoryService;
+    private final RoomRepository roomRepository;
+    private final InventoryRepository inventoryRepository;
 
 
 
@@ -48,8 +49,8 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
 
-//        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-//        hotel.setOwner(user);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
 
         hotel = hotelRepository.save(hotel);
         log.info("Created a new hotel with ID: {}", hotelDto.getId());
@@ -63,11 +64,11 @@ public class HotelServiceImpl implements HotelService {
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + id));
 
-//        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-//
-//        if (!user.equals(hotel.getOwner())) {
-//            throw new UnAuthorisedException("This user does not own this hotel with id: " + id);
-//        }
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: " + id);
+        }
 
         return modelMapper.map(hotel, HotelDto.class);
     }
@@ -79,10 +80,10 @@ public class HotelServiceImpl implements HotelService {
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + id));
 
-//        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-//        if (!user.equals(hotel.getOwner())) {
-//            throw new UnAuthorisedException("This user does not own this hotel with id: " + id);
-//        }
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: " + id);
+        }
 
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
@@ -97,16 +98,16 @@ public class HotelServiceImpl implements HotelService {
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + id));
 
-//        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-//        if (!user.equals(hotel.getOwner())) {
-//            throw new UnAuthorisedException("This user does not own this hotel with id: " + id);
-//        }
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: " + id);
+        }
 
 
-//        for (Room room : hotel.getRooms()) {
-//            inventoryService.deleteAllInventories(room);
-//            roomRepository.deleteById(room.getId());
-//        }
+        for (Room room : hotel.getRooms()) {
+            inventoryService.deleteAllInventories(room);
+            roomRepository.deleteById(room.getId());
+        }
         hotelRepository.deleteById(id);
     }
 
@@ -118,57 +119,61 @@ public class HotelServiceImpl implements HotelService {
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + hotelId));
 
-//        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-//
-//        if (!user.equals(hotel.getOwner())) {
-//            throw new UnAuthorisedException("This user does not own this hotel with id: " + hotelId);
-//        }
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: " + hotelId);
+        }
 
         hotel.setActive(true);
 
         // assuming only do it once
-//        for (Room room : hotel.getRooms()) {
-//            inventoryService.initializeRoomForAYear(room);
-//        }
+        for (Room room : hotel.getRooms()) {
+            inventoryService.initializeRoomForAYear(room);
+        }
     }
 
 
     //    public method
-//    @Override
-//    public HotelInfoDto getHotelInfoById(Long hotelId, HotelInfoRequestDto hotelInfoRequestDto) {
-//        Hotel hotel = hotelRepository
-//                .findById(hotelId)
-//                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + hotelId));
-//
-//        long daysCount = ChronoUnit.DAYS.between(hotelInfoRequestDto.getStartDate(), hotelInfoRequestDto.getEndDate()) + 1;
-//
-//        List<RoomPriceDto> roomPriceDtoList = inventoryRepository.findRoomAveragePrice(hotelId,
-//                hotelInfoRequestDto.getStartDate(), hotelInfoRequestDto.getEndDate(),
-//                hotelInfoRequestDto.getRoomsCount(), daysCount);
-//
-//        List<RoomPriceResponseDto> rooms = roomPriceDtoList.stream()
-//                .map(roomPriceDto -> {
-//                    RoomPriceResponseDto roomPriceResponseDto = modelMapper.map(roomPriceDto.getRoom(),
-//                            RoomPriceResponseDto.class);
-//                    roomPriceResponseDto.setPrice(roomPriceDto.getPrice());
-//                    return roomPriceResponseDto;
-//                })
-//                .collect(Collectors.toList());
-//
-//        return new HotelInfoDto(modelMapper.map(hotel, HotelDto.class), rooms);
-//    }
-//
-//    @Override
-//    public List<HotelDto> getAllHotels() {
-//        User user = getCurrentUser();
-//        log.info("Getting all hotels for the admin user with ID: {}", user.getId());
-//        List<Hotel> hotels = hotelRepository.findByOwner(user);
-//
-//        return hotels
-//                .stream()
-//                .map((element) -> modelMapper.map(element, HotelDto.class))
-//                .collect(Collectors.toList());
-//    }
-//
+    @Override
+    public HotelInfoDto getHotelInfoById(Long hotelId, HotelInfoRequestDto hotelInfoRequestDto) {
+        Hotel hotel = hotelRepository
+                .findById(hotelId)
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + hotelId));
 
-}
+
+        long daysCount = ChronoUnit.DAYS.between(hotelInfoRequestDto.getStartDate(), hotelInfoRequestDto.getEndDate()) + 1;
+
+        List<RoomPriceDto> roomPriceDtoList = inventoryRepository.findRoomAveragePrice(hotelId,
+                hotelInfoRequestDto.getStartDate(), hotelInfoRequestDto.getEndDate(),
+                hotelInfoRequestDto.getRoomsCount(), daysCount);
+
+        List<RoomPriceResponseDto> rooms = roomPriceDtoList.stream()
+                .map(roomPriceDto -> {
+                    RoomPriceResponseDto roomPriceResponseDto = modelMapper.map(roomPriceDto.getRoom(),
+                            RoomPriceResponseDto.class);
+                    roomPriceResponseDto.setPrice(roomPriceDto.getPrice());
+                    return roomPriceResponseDto;
+                })
+                .collect(Collectors.toList());
+
+        return new HotelInfoDto(modelMapper.map(hotel, HotelDto.class), rooms);
+    }
+
+    @Override
+    public List<HotelDto> getAllHotels() {
+        User user = getCurrentUser();
+        log.info("Getting all hotels for the admin user with ID: {}", user.getId());
+        List<Hotel> hotels = hotelRepository.findByOwner(user);
+
+        return hotels
+                .stream()
+                .map((element) -> modelMapper.map(element, HotelDto.class))
+                .collect(Collectors.toList());
+    }
+
+ }
+
+
+
+
